Added basic structure for the vm
The vm is composed of a simple array of registers
and memory that contains all of it's working
inforamtion.

For instructions, I have created variants that
allow to encapsulate properly the parameters
of each instruction to make it simple. This
logic of decoding is encapsulated in the
constants.rs file.